change point eltype

+++ src/ADT/Header/point.h
@@ -8,8 +8,8 @@
 #include "boolean.h"
 
 typedef struct { 
-	float X; /* absis   */
-	float Y; /* ordinat */
+	int X; /* absis   */
+	int Y; /* ordinat */
 } POINT;
 
 /* *** Notasi Akses: Selektor POINT *** */
@@ -18,7 +18,7 @@ typedef struct {
         
 /* *** DEFINISI PROTOTIPE PRIMITIF *** */
 /* *** Konstruktor membentuk POINT *** */
-POINT P_makePOINT (float X, float Y);
+POINT P_makePOINT (int X, int Y);
 /* Membentuk sebuah POINT dari komponen-komponennya */
 
 /* *** KELOMPOK Interaksi dengan I/O device, BACA/TULIS  *** */                                                 
@@ -62,19 +62,19 @@ POINT P_NextX (POINT P);
 /* Mengirim salinan P dengan absis ditambah satu */              
 POINT P_NextY (POINT P);
 /* Mengirim salinan P dengan ordinat ditambah satu */
-POINT P_PlusDelta (POINT P, float deltaX, float deltaY);
+POINT P_PlusDelta (POINT P, int deltaX, int deltaY);
 /* Mengirim salinan P yang absisnya adalah Absis(P) + deltaX dan ordinatnya adalah Ordinat(P) + deltaY */
 POINT P_MirrorOf (POINT P, boolean SbX);
 /* Menghasilkan salinan P yang dicerminkan terhadap salah satu sumbu */
 /* Jika SbX bernilai true, maka dicerminkan terhadap sumbu X */
 /* Jika SbX bernilai false, maka dicerminkan terhadap sumbu Y */
-float P_Jarak0 (POINT P);
+int P_Jarak0 (POINT P);
 /* Menghitung jarak P ke (0,0) */
-float P_Jarak (POINT P1, POINT P2);
+int P_Jarak (POINT P1, POINT P2);
 /* Menghitung panjang garis yang dibentuk P1 dan P2 */
 /* Perhatikanlah bahwa di sini spec fungsi kurang baik sebab menyangkut ADT Garis. */
 /* Tuliskan spec fungsi yang lebih tepat. */
-void P_Geser (POINT *P, float deltaX, float deltaY);
+void P_Geser (POINT *P, int deltaX, int deltaY);
 /* I.S. P terdefinisi */
 /* F.S. P digeser, absisnya sebesar deltaX dan ordinatnya sebesar deltaY */
 void P_GeserKeSbX (POINT *P);
